refactor: move main.go back to single class
